requirements.txt
beautifulsoup4==4.12.3
crewai==0.36.0
faiss-cpu==1.8.0
openai==1.30.3
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.36.0
tiktoken==0.7.0